requirements.txt

Flask
gunicorn
python-dotenv
flask-cors
requests
